pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lambda_hat*"]

[project]
name = "lambda-hat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "arviz~=0.22.0",
    "blackjax==1.2.5",
    "click>=8.2.1",
    "optax>=0.2.5",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "tqdm>=4.67.1",
]

[project.scripts]
lambda-hat = "lambda_hat.cli:cli"

[project.optional-dependencies]
# SLURM backend
slurm = ["submitit>=1.5"]

# Modal backend
modal = ["modal>=0.64"]

# Local / macOS / CPU (MPS or CPU-only)
cpu = [
  "jax>=0.4.28",               # pulls CPU/MPS wheels from PyPI
]

# CUDA 12.x wheels for Linux (bundled CUDA libs; only needs a recent NVIDIA driver)
cuda12 = [
  "jax[cuda12_local]>=0.4.28"
]

[dependency-groups]
dev = [
    "marimo>=0.15.2",
    "pytest>=8.4.2",
]
